+++ odoo/local-src/specific_zetes/tests/test_itemmove_parking.py
@@ -4,7 +4,7 @@ import mock
 from .. import constants
 from ..tools.domain_interface import Parameters
 from ..tools.domain_itemmove import Itemmove
-from .zetes_test_classes import DEFAULT_HEADER, ZetesParkingTest
+from .zetes_test_classes import ZetesParkingTest
 
 
 class TestItemmoveParking(ZetesParkingTest):
@@ -40,9 +40,7 @@ class TestItemmoveParking(ZetesParkingTest):
         :return:
         """
         domain = Itemmove(
-            DEFAULT_HEADER,
-            mock.MagicMock(name='Savepoint()'),
-            request_overwrite=self,
+            self._default_header(), mock.MagicMock(name='Savepoint()')
         )
 
         request_params = Parameters(domain, action='requ')
@@ -119,9 +117,7 @@ class TestItemmoveParking(ZetesParkingTest):
         self.assertEqual(pack_op.qty_done, 100)
 
         domain = Itemmove(
-            DEFAULT_HEADER,
-            mock.MagicMock(name='Savepoint()'),
-            request_overwrite=self,
+            self._default_header(), mock.MagicMock(name='Savepoint()')
         )
         request_params = Parameters(domain, action='resu')
         request_params.update(