+++ odoo/local-src/specific_zetes/tests/test_usercontext.py
@@ -4,7 +4,7 @@ import mock
 from .. import constants
 from ..tools.domain_interface import Parameters
 from ..tools.domain_usercontext import Usercontext
-from .zetes_test_classes import DEFAULT_HEADER, ZetesTest
+from .zetes_test_classes import ZetesTest
 
 
 class TestUsercontext(ZetesTest):
@@ -13,9 +13,7 @@ class TestUsercontext(ZetesTest):
 
         # Check with no current picking
         domain = Usercontext(
-            DEFAULT_HEADER,
-            mock.MagicMock(name='Savepoint()'),
-            request_overwrite=self,
+            self._default_header(), mock.MagicMock(name='Savepoint()')
         )
         request_params = Parameters(domain, action='requ')
         request_params.update({'contextType': '1'})
@@ -24,7 +22,7 @@ class TestUsercontext(ZetesTest):
         self.assertEqual(result.unitSlam, '0')
 
         # Assign the picking to the current operator
-        self.picking.operator_id = self.user.id
+        self.picking.operator_id = self.operator_user.id
         result_str = domain.requ(request_params)
         result = self.format_result(result_str)
         self.assertEqual(result.unitSlam, '1')